[138] Move view presenters
Namespace the presenters that we're using in the views to `View::` and
move them into `app/presenters/view`.

We have presenters that to present the payload for DTTP so this keeps
them separate.
